Moved AnkiConnect into anki.ts
