+++ src/anki.ts
@@ -1,0 +1,35 @@
+const ANKI_PORT: number = 8765
+
+export function request(action: string, params={}) {
+	return {action, version:6, params}
+}
+
+export function invoke(action: string, params={}) {
+    return new Promise((resolve, reject) => {
+        const xhr = new XMLHttpRequest()
+        xhr.addEventListener('error', () => reject('failed to issue request'));
+        xhr.addEventListener('load', () => {
+            try {
+                const response = JSON.parse(xhr.responseText);
+                if (Object.getOwnPropertyNames(response).length != 2) {
+                    throw 'response has an unexpected number of fields';
+                }
+                if (!response.hasOwnProperty('error')) {
+                    throw 'response is missing required error field';
+                }
+                if (!response.hasOwnProperty('result')) {
+                    throw 'response is missing required result field';
+                }
+                if (response.error) {
+                    throw response.error;
+                }
+                resolve(response.result);
+            } catch (e) {
+                reject(e);
+            }
+        });
+
+        xhr.open('POST', 'http://127.0.0.1:' + ANKI_PORT.toString());
+        xhr.send(JSON.stringify({action, version: 6, params}));
+    });
+}
